% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lookup_tables.R
\name{get_lookup_tables}
\alias{get_lookup_tables}
\title{get_lookup_tables}
\usage{
get_lookup_tables(
  mp_service_account_name = Sys.getenv("MP_SERVICE_ACCOUNT_NAME"),
  mp_service_account_secret = Sys.getenv("MP_SERVICE_ACCOUNT_TOKEN"),
  mp_project_id = Sys.getenv("MP_PROJECT_ID")
)
}
\arguments{
\item{mp_service_account_name}{The Mixpanel Service Account Name}

\item{mp_service_account_secret}{The Mixpanel Service Account Token}

\item{mp_project_id}{str The Mixpanel Project ID}
}
\value{
Tibble with information about Lookup tables
}
\description{
get_lookup_tables
}
