% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixpanelutils-package.R
\docType{package}
\name{mixpanelutils-package}
\alias{mixpanelutils}
\alias{mixpanelutils-package}
\title{mixpanelutils: Mixpanel Utils Package}
\description{
Functions to help analyzing event data in Mixpanel.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/L-Groeninger/mixpanelutils}
  \item Report bugs at \url{https://github.com/L-Groeninger/mixpanelutils/issues}
}

}
\author{
\strong{Maintainer}: Lukas Gröninger \email{lukas.groeninger@gmail.com}

}
\keyword{internal}
