% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_lookup_table.R
\name{replace_lookup_table}
\alias{replace_lookup_table}
\title{replace_lookup_table}
\usage{
replace_lookup_table(
  mp_service_account_name = Sys.getenv("MP_SERVICE_ACCOUNT_NAME"),
  mp_service_account_secret = Sys.getenv("MP_SERVICE_ACCOUNT_TOKEN"),
  mp_project_id = Sys.getenv("MP_PROJECT_ID"),
  lookup_table_id,
  table_to_replace
)
}
\arguments{
\item{mp_service_account_name}{The Mixpanel Service Account Name}

\item{mp_service_account_secret}{The Mixpanel Service Account Token}

\item{mp_project_id}{str The Mixpanel Project ID}

\item{lookup_table_id}{str The id to identify the lookup table}

\item{table_to_replace}{The dataframe to replace the lookup table}
}
\value{
Success Message if Lookup table has been replaced successfully
}
\description{
replace_lookup_table
}
